refactor(frontend): extract shared studio layout
