+++ frontend/src/app/(studio)/layout.tsx
@@ -1,0 +1,16 @@
+import { AppSidebar } from "@/components/app-sidebar";
+
+export default function StudioLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <div className="app-shell">
+      <AppSidebar />
+      <main className="workspace" aria-label="Workspace">
+        {children}
+      </main>
+    </div>
+  );
+}
